Move lander components to subdirectory
